partial helpcommand rewrite
I have re written parge parts of the help command focusing on the part for groups and normal commands. It is now a lot cleaner and makes more sense to read. I will continue on the other part of the help command in the future but for now it works as intented